refactor: move CreatedAt to Entity
