+++ src/domain/seedwork/aggregate-root.ts
@@ -1,0 +1,3 @@
+import { Entity } from '@/domain/seedwork'
+
+export abstract class AggregateRoot extends Entity { }
