+++ src/domain/seedwork/index.ts
@@ -1,0 +1,2 @@
+export * from './entity'
+export * from './aggregate-root'
